requirements.txt
accelerate==1.13.0
diffusers==0.36.0
fastapi==0.135.1
fastapi-utils==0.8.0
httpx==0.28.1
huggingface_hub==1.5.0
pillow==12.1.1
prometheus-fastapi-instrumentator==7.1.0
pydantic==2.12.5
pydantic_core==2.41.5
safetensors==0.7.0
sentencepiece==0.2.0
tokenizers==0.22.2
torch==2.10.0
torchdata==0.7.0
torchtext==0.16.0+cpu
torchvision==0.25.0
transformers==5.3.0
uvicorn==0.41.0
python-dotenv>=1.0.0